Move ipc-ready-mess-to-promise fn into separate helper
