+++ src/manage-worker-messages/helpers/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const ipcReadyMessToPromise = require('./ipc-ready-mess-to-promise')
+
+module.exports = {
+  ipcReadyMessToPromise
+}
