+++ src/manage-worker-messages/index.js
@@ -3,83 +3,33 @@
 const { app, BrowserWindow } = require('electron')
 const i18next = require('i18next')
 
-const {
-  IpcMessageError,
-  AppInitializationError
-} = require('./errors')
-const wins = require('./window-creators/windows')
-const relaunch = require('./relaunch')
+const { AppInitializationError } = require('../errors')
+const wins = require('../window-creators/windows')
+const relaunch = require('../relaunch')
 const showMessageModalDialog = require(
-  './show-message-modal-dialog'
+  '../show-message-modal-dialog'
 )
 const isMainWinAvailable = require(
-  './helpers/is-main-win-available'
+  '../helpers/is-main-win-available'
 )
-const {
-  deserializeError
-} = require('./helpers/utils')
 const showTrxTaxReportNotification = require(
-  './show-notification/show-trx-tax-report-notification'
+  '../show-notification/show-trx-tax-report-notification'
 )
 const showSyncNotification = require(
-  './show-notification/show-sync-notification'
+  '../show-notification/show-sync-notification'
 )
 const PROCESS_MESSAGES = require(
-  '../bfx-reports-framework/workers/loc.api/process.message.manager/process.messages'
+  '../../bfx-reports-framework/workers/loc.api/process.message.manager/process.messages'
 )
 const PROCESS_STATES = require(
-  '../bfx-reports-framework/workers/loc.api/process.message.manager/process.states'
+  '../../bfx-reports-framework/workers/loc.api/process.message.manager/process.states'
 )
 
+const {
+  ipcReadyMessToPromise
+} = require('./helpers')
+
 const modalDialogPromiseSet = new Set()
-
-const ipcReadyMessToPromise = (ipc) => {
-  return new Promise((resolve, reject) => {
-    try {
-      const timeout = setTimeout(() => {
-        rmHandler()
-        reject(new AppInitializationError())
-      }, 30 * 60 * 1000).unref()
-
-      const rmHandler = () => {
-        ipc.off('message', handler)
-        clearTimeout(timeout)
-      }
-      const handler = (mess) => {
-        if (
-          mess ||
-          typeof mess === 'object' ||
-          typeof mess.err === 'string'
-        ) {
-          mess.err = deserializeError(mess.err)
-        }
-
-        const { state, err } = mess ?? {}
-
-        if (typeof state !== 'string') {
-          rmHandler()
-          reject(new IpcMessageError())
-
-          return
-        }
-        if (state === 'error:app-init') {
-          rmHandler()
-          reject(err || new AppInitializationError())
-
-          return
-        }
-        if (state === 'ready:server') {
-          rmHandler()
-          resolve(mess)
-        }
-      }
-
-      ipc.on('message', handler)
-    } catch (err) {
-      reject(err)
-    }
-  })
-}
 
 const resolveModalDialogInSequence = async (asyncHandler) => {
   let resolve = () => {}
